refactor: keywords/search -> posts/search로 api 변경 (#123)
* refactor: keywords/search -> posts/search로 api 변경

* test: controller 변경에 따라 테스트 코드 수정
