+++ backend/src/test/java/site/kkokkio/domain/keyword/controller/KeywordControllerTest.java
@@ -75,83 +75,6 @@ public class KeywordControllerTest {
 	}
 
 	@Test
-	@DisplayName("키워드 조회 - 최신순 (기본)")
-	public void findKeywordTest_Default() throws Exception {
-		// Given
-		Sort sort = Sort.by("createdAt").descending();
-		PageRequest pageRequest = PageRequest.of(page, size, sort);
-
-		// createdAt 기준으로 정렬된 Mock 데이터
-		List<PostDto> sortedByCreatedAt = new ArrayList<>(postDtos);
-		sortedByCreatedAt.sort((p1, p2) -> p2.postId().compareTo(p1.postId())); // 임의의 최신순 가정
-		Page<PostDto> postDtoPage = new PageImpl<>(sortedByCreatedAt, pageRequest, postDtos.size());
-
-		when(keywordService.getPostListByKeyword(eq(keywordText), any(PageRequest.class)))
-			.thenReturn(postDtoPage);
-
-		// When & Then
-		mvc.perform(MockMvcRequestBuilders.get("/api/v1/keywords/search")
-				.param("keyword", keywordText)
-				.param("page", String.valueOf(page))
-				.param("size", String.valueOf(size))
-				.param("sort", "createdAt,desc")
-				.contentType(MediaType.APPLICATION_JSON))
-			.andExpect(MockMvcResultMatchers.status().isOk())
-			.andExpect(MockMvcResultMatchers.jsonPath("$.data.list[0].postId").value(3)) // 정렬된 Mock 데이터 확인
-			.andExpect(MockMvcResultMatchers.jsonPath("$.data.list[1].postId").value(2))
-			.andExpect(MockMvcResultMatchers.jsonPath("$.data.list[2].postId").value(1))
-			.andDo(print());
-	}
-
-	@Test
-	@DisplayName("키워드 조회 - 제목 오름차순")
-	public void findKeywordTest_TitleAsc() throws Exception {
-		// Given
-		Sort sort = Sort.by("title").ascending();
-		PageRequest pageRequest = PageRequest.of(page, size, sort);
-
-		// 제목 오름차순으로 정렬된 Mock 데이터
-		List<PostDto> sortedByTitle = new ArrayList<>(postDtos);
-		sortedByTitle.sort((p1, p2) -> p1.title().compareTo(p2.title()));
-		Page<PostDto> postDtoPage = new PageImpl<>(sortedByTitle, pageRequest, postDtos.size());
-
-		when(keywordService.getPostListByKeyword(eq(keywordText), any(PageRequest.class)))
-			.thenReturn(postDtoPage);
-
-		// When & Then
-		mvc.perform(MockMvcRequestBuilders.get("/api/v1/keywords/search")
-				.param("keyword", keywordText)
-				.param("page", String.valueOf(page))
-				.param("size", String.valueOf(size))
-				.param("sort", "title,asc")
-				.contentType(MediaType.APPLICATION_JSON))
-			.andExpect(MockMvcResultMatchers.status().isOk())
-			.andExpect(MockMvcResultMatchers.jsonPath("$.data.list[0].postId").value(1)) // 정렬된 Mock 데이터 확인
-			.andExpect(MockMvcResultMatchers.jsonPath("$.data.list[1].postId").value(2))
-			.andExpect(MockMvcResultMatchers.jsonPath("$.data.list[2].postId").value(3))
-			.andDo(print());
-	}
-
-	@Test
-	@DisplayName("키워드 조회 - 조회 실패")
-	public void findKeywordTest_Fail() throws Exception {
-		// Given
-		when(keywordService.getPostListByKeyword(eq(keywordText), any(PageRequest.class)))
-			.thenThrow(new ServiceException("404", "포스트가 존재하지 않습니다."));
-
-		// When & Then
-		mvc.perform(MockMvcRequestBuilders.get("/api/v1/keywords/search")
-				.param("keyword", keywordText)
-				.param("page", String.valueOf(page))
-				.param("size", String.valueOf(size))
-				.contentType(MediaType.APPLICATION_JSON))
-			.andExpect(MockMvcResultMatchers.status().isNotFound())
-			.andExpect(MockMvcResultMatchers.jsonPath("$.code").value("404"))
-			.andExpect(MockMvcResultMatchers.jsonPath("$.message").value("포스트가 존재하지 않습니다."))
-			.andDo(print());
-	}
-
-	@Test
 	@DisplayName("키워드 검색 소스 - 성공")
 	public void getKeywordSearchSources_Success() throws Exception {
 		// given
